+++ SiLadhida.Core/Entities/Pesanan.cs
@@ -9,6 +9,6 @@ namespace SiLadhida.Core.Entities
         public StateOrder StatusSekarang { get; set; }
 
         public List<OrderItem> Items { get; set; } = new();
-        public int TotalHarga { get; set; }
+        public decimal TotalHarga { get; set; }
     }
 }
